TEST: Refactoring ArribaManagementServiceTests
Creating a ArribaServiceBase to be used as a base class for Management and Query services
