+++ src/Arriba/Arriba.Test/Services/Management/DeleteTableForUserServiceTests.cs
